pyproject.toml
[project]
name = "entsoe-apy"
authors = [
  {name="Jonathan Berrisch", email="jonathan.berrisch@uni-due.de"},
  {name="Jie Xu", email="jie.xu@uni-due.de"}
]
version = "0.2.0"
description = "A Python library for accessing ENTSO-E Transparency Platform API endpoints"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.11.0"
keywords = ["entsoe", "energy", "electricity", "api", "transparency", "platform"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "httpx~=0.28.1",
  "loguru~=0.7.2",
  "xsdata-pydantic>=24.5",
]

[project.urls]
"Homepage" = "https://entsoe-apy.berrisch.biz/"
"Documentation" = "https://entsoe-apy.berrisch.biz/"
"Repository" = "https://github.com/BerriJ/entsoe-api-py"
"Bug Tracker" = "https://github.com/BerriJ/entsoe-api-py/issues"

[dependency-groups]
dev = ["pytest~=8.3.3", "pytest-xdist~=3.6.1", "ruff==0.12.4"]

[tool.ruff]
exclude = [
    "src/entsoe/xml_models",
    "__pycache__",
    "build",
    "*.egg-info",
]

[tool.ruff.lint]
preview = true
extend-select = ["E302", "E301", "I"]

[tool.ruff.format]
preview = true

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true
known-first-party = ["entsoe"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py", "test-*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "-v",
    "--tb=short",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
